# public/openapi.yaml
openapi: 3.0.3
info:
  title: Idir.ai Blog API
  description: |
    RESTful API for managing bilingual blog posts (English/Spanish).

    **Key Features:**
    - Clean RESTful design
    - Bilingual post support with translation groups
    - Flexible filtering and grouping
    - Media upload for cover images

    **Authentication:**
    Write operations require admin authentication via session cookie.
  version: 2.0.0
  contact:
    name: Idir Ouhab Meskine
    url: https://idir.ai
servers:
  - url: https://idir.ai/api
    description: Production server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Get a single post
      description: Get a blog post by ID
      operationId: getPost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: draft
          in: query
          description: Include drafts (requires authentication)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BlogPost'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Posts
      summary: Update a post
      description: |
        Update a blog post.

        **Permissions:**
        - Authors can update their own posts
        - Owners can update any post
        - Only Admins/Owners can change status to published
      operationId: updatePost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BlogPost'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Posts
      summary: Delete a post
      description: Delete a blog post (owner only)
      operationId: deletePost
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /media:
    post:
      tags:
        - Media
      summary: Upload media file
      description: |
        Upload an image for use as blog post cover image.

        **Restrictions:**
        - Max file size: 5MB
        - Allowed formats: JPEG, PNG, WebP, AVIF, GIF
        - Requires admin authentication
      operationId: uploadMedia
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Media uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  url:
                    type: string
                    format: uri
                    description: Public URL of uploaded media
                  path:
                    type: string
                    description: Storage path
        '400':
          description: Invalid file
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  schemas:
...
    BlogPost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        cover_image:
          type: string
          format: uri
          nullable: true
        meta_description:
          type: string
          nullable: true
        meta_keywords:
          type: array
          items:
            type: string
          nullable: true
        category:
          type: string
          enum: [insights, learnings, opinion]
        tags:
          type: array
          items:
            type: string
        language:
          type: string
          enum: [en, es]
        status:
          type: string
          enum: [draft, published]
        published_at:
          type: string
          format: date-time
          nullable: true
        translation_group_id:
          type: string
          format: uuid
          nullable: true
        view_count:
          type: integer
        read_time_minutes:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author_id:
          type: string
          format: uuid
        author_name:
          type: string
          nullable: true
...
    PostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        cover_image:
          type: string
          format: uri
        category:
          type: string
          enum: [insights, learnings, opinion]
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published]
        meta_description:
          type: string
        meta_keywords:
          type: array
          items:
            type: string
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Post not found"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session